Move Command and its subclasses into mira.command package, the rest is in mira package
