+++ src/main/java/mira/Event.java
@@ -1,4 +1,5 @@
-import java.text.ParseException;
+package mira;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
