+++ src/main/java/mira/MiraException.java
@@ -1,3 +1,5 @@
+package mira;
+
 /**
  * Represents exceptions specific to the Mira chatbot.
  */
