+++ src/main/java/mira/Savable.java
@@ -1,3 +1,5 @@
+package mira;
+
 import java.io.IOException;
 
 /**
