+++ src/main/java/mira/Storage.java
@@ -1,3 +1,5 @@
+package mira;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
