+++ src/main/java/mira/Task.java
@@ -1,3 +1,5 @@
+package mira;
+
 /**
  * Represents a task with a description and completion status.
  */
