+++ src/main/java/mira/command/DeadlineCommand.java
@@ -1,3 +1,9 @@
+package mira.command;
+
+import mira.Deadline;
+import mira.Savable;
+import mira.Storage;
+
 import java.io.IOException;
 
 /**
@@ -7,7 +13,7 @@ public class DeadlineCommand extends Command implements Savable {
     private final Deadline deadline;
 
     /**
-     * Constructs a {@code DeadlineCommand} with the specified description and due date.
+     * Constructs a {@code mira.command.DeadlineCommand} with the specified description and due date.
      *
      * @param description The description of the deadline task.
      * @param by The due date of the deadline task.
