+++ src/main/java/mira/command/DeleteCommand.java
@@ -1,10 +1,16 @@
+package mira.command;
+
+import mira.Savable;
+import mira.Storage;
+import mira.Task;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command implements Savable {
     private final int index;
 
     /**
-     * Constructs a {@code DeleteCommand} with the specified index.
+     * Constructs a {@code mira.command.DeleteCommand} with the specified index.
      *
      * @param index The index of the task.
      */
