+++ src/main/java/mira/command/EventCommand.java
@@ -1,10 +1,16 @@
+package mira.command;
+
+import mira.Event;
+import mira.Savable;
+import mira.Storage;
+
 import java.io.IOException;
 
 public class EventCommand extends Command implements Savable {
     private final Event event;
 
     /**
-     * Constructs a {@code EventCommand} with the specified description and due date.
+     * Constructs a {@code mira.command.EventCommand} with the specified description and due date.
      *
      * @param description The description of the deadline task.
      * @param from The start date of the event task.
