+++ src/main/java/mira/command/ListCommand.java
@@ -1,3 +1,5 @@
+package mira.command;
+
 /**
  * Represents a command to list all tasks in the task list.
  */
