+++ src/main/java/mira/command/MarkCommand.java
@@ -1,29 +1,34 @@
+package mira.command;
+
+import mira.Savable;
+import mira.Storage;
+
 import java.io.IOException;
 
 /**
  * Represents a command to mark a task as done in the task list.
  */
-public class UnmarkCommand extends Command implements Savable {
+public class MarkCommand extends Command implements Savable {
     private final int index;
 
     /**
-     * Constructs a {@code UnmarkCommand} with the specified index of the task to be marked.
+     * Constructs a {@code mira.command.MarkCommand} with the specified index of the task to be marked.
      *
-     * @param index The index of the task to mark as undone (1-based index).
+     * @param index The index of the task to mark as done (1-based index).
      */
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     /**
-     * Executes the command by unmarking the specified task as done.
+     * Executes the command by marking the specified task as done.
      *
-     * @return A message indicating that the task has been marked as undone.
+     * @return A message indicating that the task has been marked as done.
      */
     @Override
     public String execute() {
-        taskList.unmarkTask(index);
-        return "OK, I've marked this task as not done yet:\n  " + taskList.getTask(index);
+        taskList.markTask(index);
+        return "Nice! I've marked this task as done:\n  " + taskList.getTask(index);
     }
 
     /**
